+++ service-implementation/src/main/java/com/liaison/mailbox/service/core/processor/FileWriter.java
@@ -41,6 +41,7 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
 
 /**
  * This will place and manages the file with respect to the given location.
@@ -88,6 +89,7 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
 
                 //get payload from spectrum
                 InputStream payload = null;
+                InputStream triggerFilePayload = null;
                 try {
 
                     payload = StorageUtilities.retrievePayload(workTicket.getPayloadURI());
@@ -116,11 +118,45 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
                         //To avoid staged file entry
                         workTicket.setAdditionalContext(MailBoxConstants.FILE_EXISTS, Boolean.TRUE.toString());
                     }
+                    if (Boolean.valueOf(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_FILE_GROUP).toString())) {
 
+                    	StagedFileDAOBase dao = new StagedFileDAOBase();
+                        List<StagedFile> stagedFiles = dao.findStagedFilesByParentGlobalProcessId(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_PARENT_GPID).toString());
+                        String fileCount = workTicket.getAdditionalContext().get(MailBoxConstants.KEY_FILE_COUNT).toString();
+                        int totalCount = Integer.parseInt(fileCount.split("/")[1]);
+    				    
+                        //get payload from spectrum
+                        if (totalCount == stagedFiles.size()) {
+
+                            File triggerFile = new File(processorPayloadLocation + File.separatorChar + workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_NAME).toString());
+
+                            //write the trigger file
+                            if (StorageUtilities.getPayloadSize(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_URI).toString()) == 0) {
+                            	if (!triggerFile.exists()) {
+                                    triggerFile.createNewFile();
+                                } else {
+                                    triggerFile.delete();
+                                    triggerFile.createNewFile();
+                                }
+                            } else {
+                            	
+                            	triggerFilePayload = StorageUtilities.retrievePayload(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_URI).toString());
+                                try {
+                                	FileOutputStream outputStream = new FileOutputStream(triggerFile);
+                                    IOUtils.copy(triggerFilePayload, outputStream);
+                                } catch (IOException e) {
+                                	throw new RuntimeException(e);
+                                }
+                            }
+                       }
+                   }
                 } finally {
                     if (payload != null) {
                         payload.close();
                     }
+                    if (triggerFilePayload != null) {
+                    	triggerFilePayload.close();
+                    } 
                 }
 
                 message = (writeStatus ? "Payload written at target location : " : "File already exists at the location - ")
@@ -266,7 +302,8 @@ public class FileWriter extends AbstractProcessor implements MailBoxProcessorI {
 
 					//In-activate the old entity
 					stagedFile.setStagedFileStatus(EntityStatus.INACTIVE.name());
-	        		dao.merge(stagedFile);
+					stagedFile.setParentGlobalProcessId(workTicket.getAdditionalContext().get(MailBoxConstants.KEY_TRIGGER_FILE_PARENT_GPID).toString());
+					dao.merge(stagedFile);
                     logDuplicateStatus(stagedFile.getFileName(), stagedFile.getFilePath(), stagedFile.getGlobalProcessId(), workTicket.getGlobalProcessId());
 				}