+++ service-implementation/src/main/java/com/liaison/mailbox/service/queue/kafka/Producer.java
@@ -40,12 +40,13 @@ import static com.liaison.mailbox.MailBoxConstants.KEY_PROCESSOR_ID;
 import static com.liaison.mailbox.MailBoxConstants.PROPERTY_SKIP_KAFKA_QUEUE;
 import static com.liaison.mailbox.MailBoxConstants.URI;
 import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_PRODUCER_PREFIX;
-import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_PRODUCER_TOPIC_NAME_DEFAULT;
-import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_PRODUCER_TOPIC_NAME_LOWSECURE;
-import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_STREAM;
+import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_TOPIC_NAME_CREATE_DEFAULT;
+import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_TOPIC_NAME_DELETE_DEFAULT;
+import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_TOPIC_NAME_CREATE_LOWSECURE;
+import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_TOPIC_NAME_DELETE_LOWSECURE;
+import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KAFKA_RELAY_PRODUCER_STREAM;
 import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KEY_SERIALIZER;
 import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.KEY_SERIALIZER_DEFAULT;
-import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.SERVERS;
 import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.VALUE_SERIALIZER;
 import static com.liaison.mailbox.service.queue.kafka.QueueServiceConstants.VALUE_SERIALIZER_DEFAULT;
 
@@ -57,19 +58,22 @@ public class Producer {
     private static final Logger LOG = LogManager.getLogger(Producer.class);
     private static DecryptableConfiguration configuration = LiaisonArchaiusConfiguration.getInstance();
     private static KafkaProducer<String, String> kafkaProducer;
-    private static String TOPIC_NAME;
+    private static String TOPIC_NAME_CREATE;
+    private static String TOPIC_NAME_DELETE;
 
     private static Producer producer = null;
-    
+
     private Producer() {
         
         String deploymentType = configuration.getString(MailBoxConstants.DEPLOYMENT_TYPE, DeploymentType.RELAY.getValue());
         
         // We have to check relay and legacy relay; No need for check conveyor.
         if (DeploymentType.RELAY.getValue().equals(deploymentType)) {
-            TOPIC_NAME = configuration.getString(KAFKA_STREAM) + configuration.getString(KAFKA_PRODUCER_TOPIC_NAME_DEFAULT);
+            TOPIC_NAME_CREATE = configuration.getString(KAFKA_RELAY_PRODUCER_STREAM) + configuration.getString(KAFKA_TOPIC_NAME_CREATE_DEFAULT);
+            TOPIC_NAME_DELETE = configuration.getString(KAFKA_RELAY_PRODUCER_STREAM) + configuration.getString(KAFKA_TOPIC_NAME_DELETE_DEFAULT);
         } else {
-            TOPIC_NAME = configuration.getString(KAFKA_STREAM) + configuration.getString(KAFKA_PRODUCER_TOPIC_NAME_LOWSECURE);
+            TOPIC_NAME_CREATE = configuration.getString(KAFKA_RELAY_PRODUCER_STREAM) + configuration.getString(KAFKA_TOPIC_NAME_CREATE_LOWSECURE);
+            TOPIC_NAME_DELETE = configuration.getString(KAFKA_RELAY_PRODUCER_STREAM) + configuration.getString(KAFKA_TOPIC_NAME_DELETE_LOWSECURE);
         }
     }
     
@@ -89,8 +93,9 @@ public class Producer {
      * Send message to topic
      * 
      * @param message
+     * @param topicName
      */
-    public void produce(String message) {
+    private void produce(String message, String topicName) {
         
         if (configuration.getBoolean(PROPERTY_SKIP_KAFKA_QUEUE, false)) {
             LOG.error(" SKIP_KAFKA_QUEUE is enabled: Unable to send message to topic");
@@ -98,23 +103,25 @@ public class Producer {
         }
         
         if (MailBoxUtil.isEmpty(message)) {
-            throw new RuntimeException("Unable to send message to topic " + TOPIC_NAME + ". " + " Message is empty");
+            throw new RuntimeException("Unable to send message to topic " + topicName + ". " + " Message is empty");
         }
 
         LOG.info("MapR Streams PRODUCER message to send" + message);
 
         try {
             kafkaProducer = new KafkaProducer<>(getProperties());
-            kafkaProducer.send(new ProducerRecord<>(TOPIC_NAME, message));
+            kafkaProducer.send(new ProducerRecord<>(topicName, message));
         } catch (Exception e) {
-            throw new RuntimeException("Unable to send message to topic " + TOPIC_NAME + ". " + e.getMessage(), e);
+            throw new RuntimeException("Unable to send message to topic " + topicName + ". " + e.getMessage(), e);
         }
     }
 
     private static Properties getProperties() {
 
         Properties producerProperties = new Properties();
-        producerProperties.setProperty(SERVERS, configuration.getString(KAFKA_PRODUCER_PREFIX + SERVERS));
+        
+        // The client will make use of all servers irrespective of which servers are specified here for bootstrapping
+        // Refer here for more details : https://kafka.apache.org/documentation/
         producerProperties.setProperty(KEY_SERIALIZER, configuration.getString(KAFKA_PRODUCER_PREFIX + KEY_SERIALIZER, KEY_SERIALIZER_DEFAULT));
         producerProperties.setProperty(VALUE_SERIALIZER, configuration.getString(KAFKA_PRODUCER_PREFIX + VALUE_SERIALIZER, VALUE_SERIALIZER_DEFAULT));
         return producerProperties;
@@ -151,10 +158,11 @@ public class Producer {
      * @param message directory message dto
      */
     public void produce(KafkaMessageType kafkaMessageType, DirectoryMessageDTO message) {
+
         KafkaMessage kafkaMessage = new KafkaMessage();
         kafkaMessage.setMessageType(kafkaMessageType);
         kafkaMessage.setDirectoryMessageDTO(message);
-        produce(marshalToJSON(kafkaMessage));
+        produce(marshalToJSON(kafkaMessage), KafkaMessageType.USERACCOUNT_CREATE.equals(kafkaMessageType) ? TOPIC_NAME_CREATE : TOPIC_NAME_DELETE);
     }
 
     /**
@@ -179,7 +187,7 @@ public class Producer {
         requestObj.put(KEY_OVERWRITE, workTicket.getAdditionalContextItem(MailBoxConstants.KEY_OVERWRITE).toString().toLowerCase());
         
         kafkaMessage.setFileWriterMsg(requestObj.toString());
-        produce(marshalToJSON(kafkaMessage));
+        produce(marshalToJSON(kafkaMessage), TOPIC_NAME_CREATE);
     }
 
     /**
@@ -191,7 +199,7 @@ public class Producer {
         KafkaMessage kafkaMessage = new KafkaMessage();
         kafkaMessage.setMessageType(kafkaMessageType);
         kafkaMessage.setProcessorGuid(processorGuid);
-        produce(marshalToJSON(kafkaMessage));
+        produce(marshalToJSON(kafkaMessage), TOPIC_NAME_CREATE);
     }
     
     private String marshalToJSON(KafkaMessage kafkaMessage) {