+++ service-implementation/src/main/java/com/liaison/mailbox/service/queue/kafka/QueueServiceConstants.java
@@ -12,20 +12,20 @@ package com.liaison.mailbox.service.queue.kafka;
 
 public interface QueueServiceConstants {
 
-    String KAFKA_STREAM = "kafka.stream";
+    String KAFKA_RELAY_PRODUCER_STREAM = "kafka.relay.producer.stream";
+    String KAFKA_RELAY_CONSUMER_STREAM = "kafka.relay.consumer.stream";
+    String KAFKA_FILE_EVENT_READER_CONSUMER_STREAM= "kafka.fileeventreader.consumer.stream";
+    
+    String KAFKA_TOPIC_NAME_CREATE_DEFAULT = "kafka.topic.name.create.default";
+    String KAFKA_TOPIC_NAME_DELETE_DEFAULT = "kafka.topic.name.delete.default";
+    
+    String KAFKA_TOPIC_NAME_CREATE_LOWSECURE = "kafka.topic.name.create.lowsecure";
+    String KAFKA_TOPIC_NAME_DELETE_LOWSECURE = "kafka.topic.name.delete.lowsecure";
 
-    String KAFKA_CONSUMER_TOPIC_NAME_DEFAULT = "kafka.consumer.topic.name.default";
-    String KAFKA_PRODUCER_TOPIC_NAME_DEFAULT = "kafka.producer.topic.name.default";
-    String KAFKA_CONSUMER_TOPIC_NAME_LOWSECURE = "kafka.consumer.topic.name.lowsecure";
-    String KAFKA_PRODUCER_TOPIC_NAME_LOWSECURE = "kafka.producer.topic.name.lowsecure";
-
-    String KAFKA_CONSUMER_TOPIC_NAME_FS_EVENT = "kafka.consumer.topic.name.fs.event";
-    String KAFKA_PRODUCER_TOPIC_NAME_FS_EVENT = "kafka.producer.topic.name.fs.event";
-    String KAFKA_CONSUMER_TOPIC_NAME_FS_EVENT_LOWSECURE = "kafka.consumer.topic.name.fs.event.lowsecure";
-    String KAFKA_PRODUCER_TOPIC_NAME_FS_EVENT_LOWSECURE = "kafka.producer.topic.name.fs.event.lowsecure";
+    String KAFKA_TOPIC_NAME_FS_EVENT_DEFAULT = "kafka.topic.name.fs.event.default";
+    String KAFKA_TOPIC_NAME_FS_EVENT_LOWSECURE = "kafka.topic.name.fs.event.lowsecure";
 
     String KAFKA_CONSUMER_PREFIX = "kafka.consumer.";
-    String SERVERS = "bootstrap.servers";
     String GROUP_ID = "group.id";
     String AUTO_COMMIT = "enable.auto.commit";
     String KEY_DESERIALIZER = "key.deserializer";